+++ frontend/src/components/pages/Home/Home.tsx
@@ -3,7 +3,7 @@ import styles from "./Home.module.css";
 import { CalenderSidebar } from "../../organisms/CalenderSidebar/CalenderSidebar";
 import { UserContext } from "../../../providers/UserProvider";
 import { getMyUserData } from "../../../api/userRequest";
-import { Calendar } from "../../molecules/Calendar/Calendar";
+import { Calendar } from "../../organisms/Calendar/Calendar";
 
 export const Home: VFC = () => {
   //  グローバルstate
